Fix leftovers from DKG Engine Refactor (#437)
